Broke out DBOpenHelper to respective repository implementations
One for each of the 3 domain objects, Restaurant, Dish, Image.
